backend/pyproject.toml
[tool.poetry]
name = "imago-backend"
version = "0.1.0"
description = "FastAPI backend for fetching media"
authors = ["Jiwon Kim <jiwon.kim.096@gmail.com>"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
elasticsearch = "^8.11.0"
aiohttp = "^3.11.18"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
pydantic = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
